mvp_streamlit_temp/requirements.txt
DateTime
docxtpl
google-generativeai
huggingface-hub
numpy
pandas
psutil
PyPDF2
regex
scikit-learn
streamlit
streamlit-lottie
torch
torchvision
transformers